# model-pricing-backend/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "大模型中转平台：模型数据中心 + 进货/出货两条定价线。统一响应包 {code,message,data,requestId}；金额一律字符串传输；时间 ISO 8601 带时区。\n错误码：400 参数校验、401 未认证、403 无权限、409 幂等/状态冲突、423 主体冻结、429 限流。\n写操作（产生业务新版本/影响资金合同）必须携带 Idempotency-Key 头。\n分页：?page=1\u0026size=20 或 ?cursor=\u0026size=，统一返回 {list,total,page,size}。",
        "title": "模型管理与定价中心 API",
        "contact": {
            "name": "后端"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/api/customer/billing": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "从 customer_profile 读 credit_limit / credit_used / deposit_amount / deposit_status / billing_cycle。账单明细 bills=[] 占位（9c-① 遗留：待计费系统接入）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户门户"
                ],
                "summary": "余额 / 授信占用 / 押金状态",
                "responses": {
                    "200": {
                        "description": "code=0；data=PortalBillingView{credit_limit, credit_used, deposit_amount, deposit_status, billing_cycle, bills}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 未认证",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "code=10004 客户不存在",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/customer/home": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "balance（与 /billing 同源）+ pending_count（DRAFT/PENDING 报价数）+ unread_notifications + common_models（本等级生效价目表的 SKU 列表，**不含价格明细以外的成本/毛利字段**）。本批单接口聚合，9c-④ 登记是否拆分为并行请求。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户门户"
                ],
                "summary": "H5 首页聚合：余额 + 待处理 + 未读通知 + 常用模型",
                "responses": {
                    "200": {
                        "description": "code=0；data=PortalHomeView{balance, pending_count, unread_notifications, common_models}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 未认证",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "code=10004 客户不存在",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/customer/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "当前客户的通知（created_at DESC 分页）。标记已读接口 POST /notifications/{id}/read 登记 9c-② 遗留。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户门户"
                ],
                "summary": "通知列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data=PortalNotificationListResult{list, total, page, size}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 未认证",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/customer/price-book": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "按当前登录客户的 level_code 返回当前 EFFECTIVE 价目表（联 model_sku + 代表组件 unit_price）。**不含 floor_price / unit_cost / margin / baseline**。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户门户"
                ],
                "summary": "当前生效价目表（本等级，分币种）",
                "responses": {
                    "200": {
                        "description": "code=0；data=PortalPriceBookView{level_code, version_no, currency, items[]}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 未认证",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "code=10004 当前等级暂无生效价目表 / 客户不存在",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/customer/quotes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回当前客户的 customer_quote（按 id DESC 分页）+ customer_price_book（合同价，SourceKind=CONTRACT）联合列表。只含自己的报价（行级过滤）。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户门户"
                ],
                "summary": "我的报价与合同",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data=PortalQuoteListResult{list, total, page, size}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 未认证",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/customer/quotes/{id}/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "报价必须是 APPROVED 状态（或 special_price_status='APPROVED' 的特价单）且未过 valid_until；行级过滤 customer_id=登录客户。单事务：customer_quote.status→EFFECTIVE + INSERT customer_price_book + audit CUSTOMER_QUOTE_ACCEPTED。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "客户门户"
                ],
                "summary": "接受报价 → 转合同价（幂等）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "customer_quote.id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "幂等键",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "code=0；data=PortalAcceptResult{quote_id, new_status, contract_cnt}",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "400": {
                        "description": "code=10001 参数非法",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "401": {
                        "description": "code=10002 未认证",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "404": {
                        "description": "code=10004 报价不存在（或不是自己的报价）",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    },
                    "409": {
                        "description": "code=10005 报价非 APPROVED / 报价已过期",
                        "schema": {
                            "$ref": "#/definitions/api.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务错误码，0=成功",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "业务数据",
                    "type": "object"
                },
                "message": {
                    "description": "错误消息，成功时 \"ok\"",
                    "type": "string",
                    "example": "ok"
                },
                "requestId": {
                    "description": "请求追踪号（X-Request-Id）",
                    "type": "string",
                    "format": "uuid",
                    "example": "01929d81-..."
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer \u003ctoken\u003e，登录后获得",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}